+++ src/simmate/workflows/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from simmate.compute import SimmateWorker as Worker
 from simmate.database import connect
 
 from .core import ErrorHandler, Workflow, workflow
-from .execution import SimmateWorker as Worker